refactor: replace auth method to auth middleware
